backend/requirements.txt
polars==1.6.0
pyarrow
requests
python-dateutil
matplotlib
tqdm
pandas
meteostat
numpy
timezonefinder
pytz
feedparser
ntscraper
torch
transformers
scipy
seaborn